+++ src/main/java/com/umc/i/member/auth/model/PostAuthNumberReq.java
@@ -1,4 +1,4 @@
-package com.umc.i.member.auth;
+package com.umc.i.member.auth.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
