+++ src/main/java/com/umc/i/member/auth/model/PostAuthNumberRes.java
@@ -1,8 +1,10 @@
-package com.umc.i.member.auth;
+package com.umc.i.member.auth.model;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
+@AllArgsConstructor
 public class PostAuthNumberRes {
 
     private boolean isSuccess;
