+++ src/main/java/com/umc/i/member/login/UserSha256.java
@@ -1,4 +1,4 @@
-package com.umc.i.member;
+package com.umc.i.member.login;
 
 import java.security.MessageDigest;
